+++ sequential/s_swap_particles.h
@@ -15,7 +15,7 @@
 #define RIGHT 0x8
 
 void s_send_particle(region *des, region *src, int part_id);
-int s_test_part_pos(particle *part, int grid_size);
-void s_send_part_to_region(region *regions, int region_idx, int part_id, int region_sqt_num, int grid_size);
+int s_test_part_pos(particle *part, double grid_size);
+void s_send_part_to_region(region *regions, int region_idx, int part_id, int region_sqt_num, double grid_size);
 
 #endif //GALAXY_SIMULATION_S_SWAP_PARTICLES_H
